backend/requirements.txt
Django==4.2.4
psycopg2-binary==2.9.5
Pillow==9.4.0
django-cors-headers==3.14.0
djoser==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
django-redis==5.3.0
django-minio-backend==3.5.0
shortuuid==1.0.11
celery[redis]==5.2.7
django-mptt==0.14.0
django-mptt-admin==2.5.1
django-debug-toolbar==4.1.0
